backend/pyproject.toml
[project]
name = "nexus-backend"
version = "0.1.0"
description = "Nexus backend — FastAPI perimeter between frontend, Cognito, MongoDB, S3, Temporal, and Redis SSE"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "gunicorn>=22",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "redis>=5.0",
    "motor>=3.5",
    "aioboto3>=13",
    "temporalio>=1.8",
    "sse-starlette>=2.1",
    "prometheus-fastapi-instrumentator>=7.0",
    "structlog>=24.1",
    "aws-xray-sdk>=2.14",
    "slowapi>=0.1.9",
    "python-magic>=0.4.27",
    "python-ulid>=3.0",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mongomock-motor>=0.0.30",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nexus_backend"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
# B008: FastAPI's `Depends(...)`, `File(...)`, `Query(...)` are idiomatic in defaults.
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
